Use TextSettings to build queryURL
