+++ client/anoixo-client/src/AppSettings.ts
@@ -1,7 +1,3 @@
-export type ServerSettings = {
-  textServerUrl: string;
-};
-
-export const NLFServerSettings: ServerSettings = {
-  textServerUrl: 'http://localhost:5000/api/text/nlf',
+export const ServerSettings = {
+  apiUrl: 'http://localhost:5000/api'
 };
