+++ client/anoixo-client/src/TextContext.tsx
@@ -1,14 +1,16 @@
-import React, { useContext, ReactNode } from 'react';
+import React, { ReactNode } from 'react';
 
 enum TextName {
   NLF
 };
+
 const TextContext = React.createContext<TextName | undefined>(undefined);
 
 type TextContextProviderProps = {
   children: ReactNode;
   text: TextName;
 };
+
 const TextContextProvider: React.FC<TextContextProviderProps> = (props: TextContextProviderProps) => {
   return (
     <TextContext.Provider value={props.text}>
@@ -17,12 +19,5 @@ const TextContextProvider: React.FC<TextContextProviderProps> = (props: TextCont
   )
 };
 
-const useTextSetting = () => {
-  const context = useContext(TextContext);
-  if (context === undefined) {
-    throw new Error('useTextSetting must be used within a TextContextProvider');
-  }
-  return context;
-};
 
-export { TextName, TextContextProvider, useTextSetting };
+export { TextContext, TextName, TextContextProvider };
